Add QModel using feed forward ANN
To estimate and learn QValues we need use an ANN, has it has been shown
that they work quite well in deep learning tasks - at least in simple
ones like the one I try to solve. All the pytorch code is hidden in the
QModel class, so the actual RL algorithm is independent of the tensor
framework used, making future switches easier.