Make stochastic tests more stable
I use a custom pytest extensions to record multiple results of training
runs and provide an average measure. This avoids the test failing in
the case where I get by chance the agent doesn't learn a good policy.
Additionally I mark some tests as 'flaky' where it is sufficient if they
run again as the expected results are usually very close and only
sometimes it happens that they are off by a small margin.