Implement train CLI
The CLI takes a config file with all training parameters to train an
agent. This not only makes usage quite simple, by keeping the config
files they automatically also serve as documentation of experiments
making it easy to reference them later if necessary. Also I've added
tests for the CLI, to make sure no subtle bugs are introduced when
training an agent - for instance as happened - forgetting to update the
observation object.