Implement continuous unity gym adapter
The next project is about a continuous control tasks, hence it was
necessary to extend the unity adapter to allow for continuous control.
Additionally the cli has changed to be able to select the algorithm to
be used, so I can select one suited for continuous control. Finally a
new verb 'exploration' has been introduced to be able to look at new
environments and get a feel for how they work. It also served as a first
test bench for the new continuous adaptation.