Implement actor and critic models
For the DDPG algorithm I need actor and critic models. These are ever so
slightly different from the ones used before. Hence a new interface has
been developed. The current solution is workable but still needs some
thought from a software engineering point of view, as the heavy usage
of inheritance indicates a possible design smell. However, for now and
to complete the assignment it has been deemed enough.