Restructure README.md and Report.md
The requirements of the udacity projects are that README and Report
files are located in the root of the repository. However, to be able to
reuse explanations and to keep detailed reports nicely structured I link
to project specific README and Report files from the root files. This
way the files are not overcrowded with information of all the projects
and I can nicely structure information that is common to all projects.