Improve exploration logic of MADDPG agent
By allowing preheat steps which are completely random the agent can
observe a more diverse set of states. Otherwise it wouldn't see that
much because it tends to crash early on into a local minimum. With these
trajectories in the replay buffer it seems to be better abel to get out
of it again. Additionally I've added static noise to the training. This
should also help in exploring and getting the agent out of local minima.
For now I just use a fixed value.