Implement keyboard interrupt and snapshots
I want to keep the state when of the agent when I interrupt training.
Often because it already reached a very high value. Also I wanted a way
of investigating agent behaviour during training to get a better feel
for why performance crashes early on. Additionally, to not loose well
trained agents I added a snapshot system.