Fix NDDPG tests and coverage report
With the new agent setup the basic learning tests failed, because it now
required different observation, action and reward shapes. Also the
multiple inheritance approach didn't work that well, because of the
order each gym wrapper was invoked. Also added a .coveragerc file to
make the coverage reports more accurate and to exclude main.py as it is
tested when using it directly.