Implement PriorityReplayBuffer
With this buffer the agent can sample its replay memory more efficiently
prioritizing experiences with high TD-Error. The TD-Error is a good
indicator as it intuitively resembles how "surprised" the agent was to
see that state. It makes sense to focus on more surprising states to
learn from when training.